fix: refactor service
