+++ src/pages/board/outlets/financialReporting/ToDo/StaffModal/index.tsx
@@ -1,7 +1,6 @@
-import { useState, useEffect, useContext } from "react";
 import { createPortal } from "react-dom";
 import { MdClear } from "react-icons/md";
-import { useNavigate, useParams } from "react-router-dom";
+
 import { Formik, Form } from "formik";
 import * as Yup from "yup";
 import {
@@ -13,21 +12,12 @@ import {
   Select,
   Blanket,
   Button,
-  useFlag,
 } from "@inubekit/inubekit";
 
-import { getCommercialManagerAndAnalyst } from "@services/commercialManagerAndAnalyst";
-import {
-  ICommercialManagerAndAnalyst,
-  ICreditRequests,
-} from "@pages/SubmitCreditApplication/types";
-
 import { StyledModal, StyledContainerClose } from "./styles";
-import { saveCredit } from "./utils";
 
-import { AppContext } from "@context/AppContext";
 import { IToDo } from "@services/types";
-import { textFlags } from "@config/pages/staffModal/addFlag";
+import { ICommercialManagerAndAnalyst } from "@pages/SubmitCreditApplication/types";
 
 export interface StaffModalProps {
   commercialManager: string;
@@ -37,153 +27,46 @@ export interface StaffModalProps {
   onSubmit?: (values: { commercialManager: string; analyst: string }) => void;
   onCloseModal?: () => void;
   taskData?: IToDo | null;
+  handleCommercialManagerChange: (
+    name: string,
+    value: string,
+    setFieldValue: (field: string, value: string) => void
+  ) => void;
+  handleAnalystChange: (
+    name: string,
+    value: string,
+    setFieldValue: (field: string, value: string) => void
+  ) => void;
+  handleCreditRequests: () => void;
+  loading: boolean;
+  accountManagerList: ICommercialManagerAndAnalyst[];
+  analystList: { staffId: string; staffName: string }[];
 }
 
 export function StaffModal(props: StaffModalProps) {
-  const { portalId = "portal", onSubmit, onCloseModal, taskData } = props;
-  const [analystList, setAnalystList] = useState<
-    ICommercialManagerAndAnalyst[]
-  >([]);
-  const [accountManagerList, setAccountManagerList] = useState<
-    ICommercialManagerAndAnalyst[]
-  >([]);
-  const [loading, setLoading] = useState(true);
-  const [selectedCommercialManager, setSelectedCommercialManager] =
-    useState<ICommercialManagerAndAnalyst | null>(null);
-  const [selectedAnalyst, setSelectedAnalyst] =
-    useState<ICommercialManagerAndAnalyst | null>(null);
+  const {
+    portalId = "portal",
+    onSubmit,
+    onCloseModal,
+    handleCommercialManagerChange,
+    handleAnalystChange,
+    handleCreditRequests,
+    loading,
+    accountManagerList,
+    analystList,
+  } = props;
+
   const isMobile = useMediaQuery("(max-width: 700px)");
-  const [showModal, setShowModal] = useState(false);
+
   const node = document.getElementById(portalId);
   const validationSchema = Yup.object().shape({
     commercialManager: Yup.string(),
     analyst: Yup.string(),
   });
-  const { id } = useParams();
-  const navigate = useNavigate();
 
   if (!node) {
     throw new Error("The portal node is not defined.");
   }
-
-  const { businessUnitSigla, eventData } = useContext(AppContext);
-  const { userAccount } =
-    typeof eventData === "string" ? JSON.parse(eventData).user : eventData.user;
-  const businessUnitPublicCode: string =
-    JSON.parse(businessUnitSigla).businessUnitPublicCode;
-
-  const handleCommercialManagerChange = (
-    name: string,
-    value: string,
-    setFieldValue: (field: string, value: string) => void
-  ) => {
-    setFieldValue(name, value);
-    const selectedManager = accountManagerList.find(
-      (manager) => manager.staffName === value
-    );
-    if (selectedManager) {
-      setSelectedCommercialManager(selectedManager);
-    }
-  };
-
-  const handleAnalystChange = (
-    name: string,
-    value: string,
-    setFieldValue: (field: string, value: string) => void
-  ) => {
-    setFieldValue(name, value);
-    const selected = analystList.find((analyst) => analyst.staffName === value);
-    if (selected) {
-      setSelectedAnalyst(selected);
-    }
-  };
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setLoading(true);
-      try {
-        const [accountManagers, analysts] = await Promise.all([
-          getCommercialManagerAndAnalyst("CredicarAccountManager", "Selsa"),
-          getCommercialManagerAndAnalyst("CredicarAnalyst", "Selsa"),
-        ]);
-        setAccountManagerList(accountManagers);
-        setAnalystList(analysts);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  const { addFlag } = useFlag();
-  const buildCreditRequest = (
-    role: string,
-    user: ICommercialManagerAndAnalyst | null
-  ): ICreditRequests | null => {
-    if (!user) return null;
-
-    return {
-      creditRequestId: taskData?.creditRequestId || "",
-      executed_task: taskData?.taskToBeDone || "",
-      execution_date: new Date().toISOString().split("T")[0],
-      identificationNumber: user.identificationDocumentNumber || "",
-      identificationType: "C",
-      role: role,
-      transactionOperation: "Insert",
-      userId: user.staffId || "",
-      userName: user.staffName || "",
-      justification: "Justificacion",
-      creditRequestCode: "",
-    };
-  };
-
-  const handleCreditRequests = async () => {
-    const managerRequest = buildCreditRequest(
-      "CredicarAccountManager".substring(0, 20),
-      selectedCommercialManager
-    );
-
-    const analystRequest = buildCreditRequest(
-      "CredicarAnalyst",
-      selectedAnalyst
-    );
-
-    try {
-      if (managerRequest) {
-        await saveCredit(businessUnitPublicCode, managerRequest, userAccount);
-      }
-
-      if (analystRequest) {
-        await saveCredit(businessUnitPublicCode, analystRequest, userAccount);
-      }
-
-      addFlag({
-        title: textFlags.titleSuccess,
-        description: textFlags.descriptionSuccess,
-        appearance: "success",
-        duration: 5000,
-      });
-    } catch (error) {
-      addFlag({
-        title: textFlags.titleError,
-        description: textFlags.descriptionError,
-        appearance: "danger",
-        duration: 5000,
-      });
-    } finally {
-      handleToggleModal();
-      setTimeout(() => {
-        navigate(`/extended-card/${id}`);
-      }, 6000);
-    }
-  };
-
-  const handleToggleModal = () => {
-    setShowModal(!showModal);
-  };
 
   const options = {
     commercialManager: accountManagerList.map((official) => ({
